Preferences: Move selection of Python interpreter to its own page
